moved BuildableIonCubePedestal and GenericPrecursorDecoration to the API namespace
