+++ ArchitectsLibrary/ArchitectsLibrary.csproj
@@ -121,7 +121,9 @@
     </Reference>
   </ItemGroup>
   <ItemGroup>
+    <Compile Include="API\BuildableIonCubePedestal.cs" />
     <Compile Include="API\EggPrefab.cs" />
+    <Compile Include="API\GenericPrecursorDecoration.cs" />
     <Compile Include="API\HolographicPoster.cs" />
     <Compile Include="API\PrecursorFabricatorService.cs" />
     <Compile Include="API\DisplayCaseServices.cs" />
@@ -130,7 +132,6 @@
     <Compile Include="Buildables\BuildableColumnSmall.cs" />
     <Compile Include="Buildables\BuildableElectricubePedestal.cs" />
     <Compile Include="Buildables\BuildableItemPedestal.cs" />
-    <Compile Include="Buildables\BuildableIonCubePedestal.cs" />
     <Compile Include="Buildables\BuildablePedestal.cs" />
     <Compile Include="Buildables\BuildablePedestalLarge.cs" />
     <Compile Include="Buildables\BuildableRedIonCubePedestal.cs" />
@@ -150,7 +151,6 @@
     <Compile Include="Buildables\BuildablePlatform.cs" />
     <Compile Include="Buildables\BuildableColumn.cs" />
     <Compile Include="Buildables\BuildableWarper.cs" />
-    <Compile Include="Buildables\GenericPrecursorDecoration.cs" />
     <Compile Include="CreatorKit\Mono\FilterableList.cs" />
     <Compile Include="CreatorKit\Mono\FilterableListItem.cs" />
     <Compile Include="CreatorKit\Mono\GenericButton.cs" />
