+++ ArchitectsLibrary/Buildables/BuildableInfoPanel.cs
@@ -1,4 +1,5 @@
-﻿using UnityEngine;
+﻿using ArchitectsLibrary.API;
+using UnityEngine;
 
 namespace ArchitectsLibrary.Buildables
 {
