+++ ArchitectsLibrary/Buildables/BuildableMicroscope.cs
@@ -2,6 +2,7 @@
 using SMLHelper.V2.Crafting;
 using System.Collections;
 using System.Collections.Generic;
+using ArchitectsLibrary.API;
 using ArchitectsLibrary.Handlers;
 
 namespace ArchitectsLibrary.Buildables
