+++ ArchitectsLibrary/Buildables/BuildableRedIonCubePedestal.cs
@@ -1,6 +1,7 @@
 ﻿using ArchitectsLibrary.Handlers;
 using SMLHelper.V2.Crafting;
 using System.Collections.Generic;
+using ArchitectsLibrary.API;
 
 namespace ArchitectsLibrary.Buildables
 {
