+++ ArchitectsLibrary/Buildables/BuildableSonicDeterrent.cs
@@ -1,6 +1,7 @@
 ﻿using SMLHelper.V2.Crafting;
 using UnityEngine;
 using System.Collections.Generic;
+using ArchitectsLibrary.API;
 using ArchitectsLibrary.Handlers;
 using ArchitectsLibrary.MonoBehaviours;
